Rename module do_action to action
